# docs/api/api.yaml
swagger: '2.0'
info:
  title: Cello API Engine Service
  description: "\n    This is swagger docs for Cello API engine.\n    "
  version: v1
basePath: /engine
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
security:
  - JWT: []
paths:
...
  /organizations/{id}/certificates:
    post:
      operationId: organizations_request_certificate
      summary: Request Certificate
      description: Request certificate
      parameters: []
      responses:
        '201':
          description: ''
        '400':
          description: ''
          schema:
            $ref: '#/definitions/BadResponse'
        '401':
          description: Permission denied
        '500':
          description: Internal Error
        '403':
          description: Authentication credentials were not provided.
      tags:
        - organizations
    parameters:
      - name: id
        in: path
        required: true
        type: string
  /organizations/{id}/users:
    get:
      operationId: organizations_users_read
      summary: List users
      description: List users in Organization
      parameters:
        - name: username
          in: query
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          required: true
          type: string
          pattern: ^[\w.@+-]+$
          maxLength: 150
          minLength: 1
        - name: page
          in: query
          description: Page of filter
          required: false
          type: integer
          default: 1
          minimum: 1
        - name: per_page
          in: query
          description: Per Page of filter
          required: false
          type: integer
          default: 10
          maximum: 100
          minimum: 1
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/UserList'
        '400':
          description: ''
          schema:
            $ref: '#/definitions/BadResponse'
        '401':
          description: Permission denied
        '500':
          description: Internal Error
        '403':
          description: Authentication credentials were not provided.
      tags:
        - organizations
    post:
      operationId: organizations_users_create
      summary: Add User
      description: Add user into Organization
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/UserID'
      responses:
        '202':
          description: Accepted
        '400':
          description: ''
          schema:
            $ref: '#/definitions/BadResponse'
        '401':
          description: Permission denied
        '500':
          description: Internal Error
        '403':
          description: Authentication credentials were not provided.
      tags:
        - organizations
    parameters:
      - name: id
        in: path
        required: true
        type: string
...
  /users:
    get:
      operationId: users_list
      summary: List Users
      description: List user through query parameter
      parameters: []
      responses:
        '200':
          description: ''
        '400':
          description: ''
          schema:
            $ref: '#/definitions/BadResponse'
        '401':
          description: Permission denied
        '500':
          description: Internal Error
        '403':
          description: Authentication credentials were not provided.
      tags:
        - users
    post:
      operationId: users_create
      summary: Create User
      description: Create new user
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/UserCreateBody'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/UserID'
        '400':
          description: ''
          schema:
            $ref: '#/definitions/BadResponse'
        '401':
          description: Permission denied
        '500':
          description: Internal Error
        '403':
          description: Authentication credentials were not provided.
      tags:
        - users
    parameters: []
...
  /users/{id}/attributes:
    get:
      operationId: users_attributes_read
      summary: Get User Attributes
      description: Get attributes of user
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - users
    post:
      operationId: users_attributes_create
      summary: Create Attributes
      description: Create attribute for user
      parameters: []
      responses:
        '201':
          description: ''
      tags:
        - users
    put:
      operationId: users_attributes_update
      summary: Update Attribute
      description: Update attribute of user
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - users
    delete:
      operationId: users_attributes_delete
      summary: Delete Attribute
      description: Delete attribute of user
      parameters: []
      responses:
        '204':
          description: ''
      tags:
        - users
    parameters:
      - name: id
        in: path
        required: true
        type: string
  /users/{id}/password:
    post:
      operationId: users_password
      summary: Update/Reset Password
      description: Update/Reset password for user
      parameters: []
      responses:
        '201':
          description: ''
        '400':
          description: ''
          schema:
            $ref: '#/definitions/BadResponse'
        '401':
          description: Permission denied
        '500':
          description: Internal Error
        '403':
          description: Authentication credentials were not provided.
      tags:
        - users
    parameters:
      - name: id
        in: path
        required: true
        type: string
definitions:
...
  BadResponse:
    required:
      - code
    type: object
    properties:
      code:
        title: Code
        description: "\n        Error Codes:\n        \n            20000: Unknown\
          \ Error.\n            \n            20001: Validation parameter error.\n\
          \            \n            20002: Parse error.\n            \n         \
          \   20003: Resource is inuse.\n            \n            20004: Resource\
          \ already exists.\n            \n            20005: Request Resource Not\
          \ found.\n            \n            20006: Permission Error.\n         \
          \   \n            20007: Custom Error.\n            \n            20008:\
          \ Have no available resource.\n            "
        type: integer
      detail:
        title: Detail
        description: Error Messages
        type: string
...
  UserInfo:
    description: Users list
    required:
      - id
      - username
    type: object
    properties:
      id:
        title: Id
        description: ID of user
        type: string
        format: uuid
      username:
        title: Username
        description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
          only.
        type: string
        pattern: ^[\w.@+-]+$
        maxLength: 150
        minLength: 1
      role:
        title: Role
        type: string
        enum:
          - 0
          - 1
          - 2
  UserList:
    required:
      - total
      - data
    type: object
    properties:
      total:
        title: Total
        description: Total number of users
        type: integer
      data:
        description: Users list
        type: array
        items:
          $ref: '#/definitions/UserInfo'
  UserID:
    required:
      - id
    type: object
    properties:
      id:
        title: Id
        description: ID of user
        type: string
        format: uuid
  UserCreateBody:
    required:
      - username
      - role
      - password
      - email
    type: object
    properties:
      username:
        title: Username
        description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
          only.
        type: string
        pattern: ^[\w.@+-]+$
        maxLength: 150
        minLength: 1
      role:
        title: Role
        description: "\n        User roles:\n            administrator\n         \
          \   \n            operator\n            \n            user\n            "
        type: string
        enum:
          - administrator
          - operator
          - user
      organization:
        title: Organization
        type: string
        format: uuid
        x-nullable: true
      password:
        title: Password
        type: string
        maxLength: 128
        minLength: 1
      email:
        title: Email address
        type: string
        format: email
        maxLength: 254